Present a clear interface to the resin.io SDK component
The resin.io platform has an interface, which consists of functions such
as download device type OS, create an application, etc.

This interface is implemented by the Resin SDK. In the future, we will
implement the same interface using the Resin CLI, and even the Dashboard
by using a browser driver technology.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>