Add a device-type component with a default implementation
A "device-type" component implements a "provision" function that given a
writer, an image, and some options:

- Flashes the image (in whatever way that makes)
- Provisions the device (in whatever way that makes sense)

After the provision function runs, we can expect a device to be up and
running.

We provide a default implementation that writes the image to a disk, and
waits for the test runner person to manually remove the SD Card and
complete the provisioning process.

This commit also introduces `utils.requireComponent()`, which will try
to find a proper implementation, or otherwise default to `default.js`.

Change-Type: patch
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>