Define OS concept
Create OS component which provides the required configuration to run a specific OS.
This class will be closely related (coupled) to the worker class. The flash method in the
worker class will use the OS class to configure the image.
The two interfaces should be kept in sync.
Because of the move of the download function we changed the way we retrive the filename.

Change-type: patch
Signed-off-by: Theodor Gherzan <theodor@resin.io>