Move from ava to node-tap
The use of ava is becoming more of a hassle than anything.
Ava is taking a lot of the freedom we have away and we had to
push back on this until ava cracked.
Node-tap seems to be a much better fit for this project as it is light-weight and
highly customizable.

Change-type: patch
Signed-off-by: Theodor Gherzan <theodor@resin.io>