Re-work core test tree representation
In order to know all test requirments when we discover
them we need to maintain the whole tree in memory.
We need this to be able to specify contracts per test check weather we satisfy them
before we run the tests. This also cleans up the test API.

Change-log: minor
Signed-off-by: Theodor Gherzan <theodor@balena.io>