Stub the proxy endpoint as glider has been removed from the worker
When we removed the dependency on glider in the proxy tests, we also
removed glider from the worker. However, older versions of the tests
from ESR branches would still call this endpoint and expect a sane
response.

Fortunately, the old proxy tests were invalid and would always pass
even if glider wasn't working. ICMP traffic does not use the proxy.
So, we can just pretend to start a glider instance so the old tests
continue to run, even with an invalid result.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>