workers: qemu: make kvm optional
KVM requires support from the CPU, BIOS/firmware, and hypervisor to all
be enabled. In the case where support is missing or disabled in any of
these places, KVM cannot be used. If the device node is missing,
attempting to map it in the compose file will fail, and the worker
cannot be started.

Instead, set a device cgroup policy to allow our container to create and
access the required device nodes at runtime if KVM is supported. This
allows a fallback to software emulation when KVM is not available. It
also allows for software emulation of targets not matching the host
architecture.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>