core: refactor Teardown as TaskQueue
Refactor Teardown class as TaskQueue to facilitate registering tasks for
setup as well as teardown. A separate Teardown object inherits from
TaskQueue and retains the signal handling from the original object.

This is a precursor to separating the forking behavior from the suite
object in preparation for having a standalone runner for virtualized
workers. For example, we can move the removeDownloads hook from the
Suite object to a separate SuiteSubprocess object that is specifically
for handling suites not running on a the same physical machine.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>