core: move suite parameters to constructor
Previously, suite parameters were read from a config file exclusively,
necessitating the creation of that config file to initialize a Suite
object.

Move the required parameters to the constructor of the Suite object to
facilitate initializing this object using other configurations, such as
command line arguments.

The suiteSubprocess module continues to initialize the Suite in the same
way as before. However, the 'interactive' property is disabled, as the
meta-balena suite appears to ignore this anyway.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>