core: test: remove intermediate suite object
The Test object is constructed with a Suite instance, which is
accessible from inside tests using 'this.suite'. However, another object
is created using only certain properties from the input suite object.

Remove the intermediate object so that all properties from the suite
object are exposed in the test code.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>