core: balenaos: use temp file for image path
The BalenaOS object has a fetch() method that unpacks a compressed disk
image from an input path to a destination path, called "path". The
destination is currently pulled from a config file.

We don't need to retain the extracted image, so switch to a temp file
path, which makes this bit of configuration unnecessary.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>