core: create WorkerClient object with address param
The Worker object retrieves the address of the Worker API to
communicate with from a config file. Rename the Worker object to
WorkerClient, to indicate that it interfaces with the API of the Worker,
and add the workerAddress parameter to the constructor.

Export a Worker object that inherits from WorkerClient, and sets the
workerAddress from a config as before.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>