workers: qemu: ignore stdio without debug
When debug is disabled, QEMU's stdio is piped to the parent process. If
that stream isn't read, it can fill and cause the child process to block
waiting for the pipe buffer to accept more data.

Override the default and ignore stdio when debug is not enabled to
prevent this.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>