core: worker: add retryOptions to executeCommand methods
The executeCommand family of methods default to retrying on failure. In
some cases, such as in the ssh-auth test in the cloud test suite, we
expect failures to happen, and want them to be raised immediately. Other
situations might demand adjusting the number of retries and interval to
fit specific tests. Add a retryOptions object to these methods to allow
for this behavior to be configured.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>