suite: add installerForceMigration configuration
Read the suite configuration for the force migration flag which configures
the flasher image to force an installer migration.

With the current QEMU setting, `installerForceMigration` alone is not
enough and the QEMU_INTERNAL_STORAGE also needs to be set to false in
the worker's environment so only a single external disk is attached to
the emulator.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>