balenaos: remove split image input/path distinction
Previously, images used for testing were gzipped, and would be extracted
to a temporary location before flashing and executing tests. This
necessitated gzipping the image before using it, and introduced
unnecessary complexity and runtime costs. Remove it in favor of a single
input image path to an uncompressed image.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>